133 retry play around to test zero downtime

+++ omnistrate_tests/test_cluster_replicas.py
@@ -3,6 +3,7 @@ import signal
 from random import randbytes
 from pathlib import Path
 import threading
+
 file = Path(__file__).resolve()
 parent, root = file.parent, file.parents[1]
 sys.path.append(str(root))
@@ -50,7 +51,7 @@ parser.add_argument("--aof-config", required=False, default="always")
 parser.add_argument("--host-count", required=False, default="6")
 parser.add_argument("--cluster-replicas", required=False, default="1")
 parser.add_argument("--shards", required=False, default="3")
-
+parser.add_argument("--persist-instance-on-fail",action="store_true")
 parser.add_argument("--ensure-mz-distribution", action="store_true")
 
 parser.set_defaults(tls=False)
@@ -66,8 +67,9 @@ def signal_handler(sig, frame):
     sys.exit(0)
 
 
-signal.signal(signal.SIGINT, signal_handler)
-signal.signal(signal.SIGTERM, signal_handler)
+if not args.persist_instance_on_fail:
+    signal.signal(signal.SIGTERM, signal_handler)
+    signal.signal(signal.SIGINT, signal_handler)
 
 current_host_count = int(args.host_count)
 current_replicas_count = int(args.cluster_replicas)
@@ -156,7 +158,8 @@ def test_cluster_replicas():
 
     except Exception as e:
         logging.exception(e)
-        instance.delete(False)
+        if not args.persist_instance_on_fail:
+            instance.delete(False)
         raise e
 
     # Delete instance
@@ -324,7 +327,7 @@ def test_zero_downtime(
     instance: OmnistrateFleetInstance,
     ssl=False,
 ):
-    """This function should test the ability to read and write while a memory update happens"""
+    """This function should test the ability to read and write while testing cluster replicas"""
     try:
         db = instance.create_connection(ssl=ssl, force_reconnect=True)
 
@@ -334,13 +337,11 @@ def test_zero_downtime(
             # Write some data to the DB
             graph.query("CREATE (n:Person {name: 'Alice'})")
             graph.ro_query("MATCH (n:Person {name: 'Alice'}) RETURN n")
-
             time.sleep(3)
     except Exception as e:
         logging.exception(e)
         error_signal.set()
         raise e
-    
 
 if __name__ == "__main__":
     test_cluster_replicas()
