+++ omnistrate_tests/test_standalone.py
@@ -46,7 +46,7 @@ parser.add_argument("--storage-size", required=False, default="30")
 parser.add_argument("--tls", action="store_true")
 parser.add_argument("--rdb-config", required=False, default="medium")
 parser.add_argument("--aof-config", required=False, default="always")
-
+parser.add_argument("--persist-instance-on-fail",action="store_true")
 parser.add_argument("--custom-network", required=False)
 
 parser.set_defaults(tls=False)
@@ -62,8 +62,9 @@ def signal_handler(sig, frame):
     sys.exit(0)
 
 
-signal.signal(signal.SIGINT, signal_handler)
-signal.signal(signal.SIGTERM, signal_handler)
+if not args.persist_instance_on_fail:
+    signal.signal(signal.SIGTERM, signal_handler)
+    signal.signal(signal.SIGINT, signal_handler)
 
 
 def test_standalone():
@@ -130,7 +131,8 @@ def test_standalone():
         test_stop_start(instance)
     except Exception as e:
         logging.exception(e)
-        instance.delete(network is not None)
+        if not args.persist_instance_on_fail:
+            instance.delete(network is not None)
         raise e
 
     # Delete instance
